Move test fixtures for parametersetdb to own file
